Update acorn to ^3.1.0
First, the ES5 fallback in `reallyParse` is now unnecessary because the
bug is now fixed in acorn.

Second, a bug-fix in the AST walker made the walker now go through
CatchClause as well, changing the number of parents before the `param`
property while making it possible to attach the locals on the entire
CatchClause, which is a cleaner solution.